rename Box index file to Box
